+++ src/main/java/com/utp/integradorspringboot/repositories/VehiculoRepository.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface VehiculoRepository extends JpaRepository<Vehiculo, Long> {
+public interface VehiculoRepository extends JpaRepository<Vehiculo, Integer> {
+    // Puedes agregar métodos personalizados aquí si lo necesitas
     List<Vehiculo> findByPlacaContainingIgnoreCaseOrMarcaContainingIgnoreCaseOrApellidosContainingIgnoreCase(String placa, String marca, String apellidos);
 }
